fix(process): preserve optional extinction waits
