+++ src/process/supervisor/supervisor.test-support.ts
@@ -1,0 +1,85 @@
+import { vi } from "vitest";
+import type { ProcessSupervisor, SpawnInput, SpawnProcessAdapter } from "./types.js";
+
+type ChildSpawnOptions = Omit<Extract<SpawnInput, { mode: "child" }>, "backendId" | "mode">;
+
+export type StubChildAdapter = SpawnProcessAdapter<NodeJS.Signals | null> & {
+  emitStdout: (chunk: string) => void;
+  emitStderr: (chunk: string) => void;
+  settle: (code: number | null, signal?: NodeJS.Signals | null) => void;
+  killMock: ReturnType<typeof vi.fn>;
+  disposeMock: ReturnType<typeof vi.fn>;
+};
+
+export function createWriteStdoutArgv(output: string): string[] {
+  if (process.platform === "win32") {
+    return [process.execPath, "-e", `process.stdout.write(${JSON.stringify(output)})`];
+  }
+  return ["/usr/bin/printf", "%s", output];
+}
+
+export function createSilentIdleArgv(): string[] {
+  return [process.execPath, "-e", "setInterval(() => {}, 1_000)"];
+}
+
+export function createStubChildAdapter(options?: {
+  pid?: number;
+  onKill?: (signal: NodeJS.Signals | undefined, adapter: StubChildAdapter) => void;
+}): StubChildAdapter {
+  const stdoutListeners: Array<(chunk: string) => void> = [];
+  const stderrListeners: Array<(chunk: string) => void> = [];
+  let resolveWait:
+    | ((value: { code: number | null; signal: NodeJS.Signals | null }) => void)
+    | null = null;
+  const waitPromise = new Promise<{ code: number | null; signal: NodeJS.Signals | null }>(
+    (resolve) => {
+      resolveWait = resolve;
+    },
+  );
+  const killMock = vi.fn();
+  const disposeMock = vi.fn();
+  const adapter: StubChildAdapter = {
+    pid: options?.pid ?? 1234,
+    stdin: undefined,
+    onStdout: (listener) => {
+      stdoutListeners.push(listener);
+    },
+    onStderr: (listener) => {
+      stderrListeners.push(listener);
+    },
+    wait: async () => await waitPromise,
+    kill: (signal) => {
+      killMock(signal);
+      options?.onKill?.(signal, adapter);
+    },
+    dispose: () => {
+      disposeMock();
+    },
+    emitStdout: (chunk) => {
+      for (const listener of stdoutListeners) {
+        listener(chunk);
+      }
+    },
+    emitStderr: (chunk) => {
+      for (const listener of stderrListeners) {
+        listener(chunk);
+      }
+    },
+    settle: (code, signal = null) => {
+      resolveWait?.({ code, signal });
+      resolveWait = null;
+    },
+    killMock,
+    disposeMock,
+  };
+
+  return adapter;
+}
+
+export async function spawnChild(supervisor: ProcessSupervisor, options: ChildSpawnOptions) {
+  return supervisor.spawn({
+    ...options,
+    backendId: "test",
+    mode: "child",
+  });
+}
